Change calculation of "unique frames" and add a unique frames metric to dash
Our previous calculation used the norm of the difference between two frames
to measure frame difference, but this didn't work too well: HDMI capture
is somewhat "noisy" (for whatever reason) and thus has many individual
pixels that are very slightly different from the previous frame. Correct
things so we only consider the frame difference based on pixels which are
qualitatively different (i.e. > 5 RGB components apart).

Also, add a number of unique frames counter to the dashboard. This may
be a more useful measure of the fluidness of an animation than FPS
(which can be deceptively low during times when nothing is happening on
a page).