Change way scrolling works to get around problem with monkeyrunner
Monkeyrunner's network mode doesn't add touch down times to the event
it sends, which causes problems when we're trying to scroll: Android's
gesture recognizer thinks we're doing a long tap, which can cause the
"save image" dialog to pop up.

Let's do the same thing we did with double tap events: upload a script
with touch down times to the device, and get the monkey tool to execute
it standalone.