Bug 762573 - Add basics of a "startup time test"
What this does is load a specific test case, then exit when it's done
loading. There's a number of different metrics you could calculate with
this, though currently the only one we do is "stable frame" -- an
indication of which point in the capture we've reached a stable state
(i.e. little to no pixel changes)