Bug 917404 - Consolidate and clarify capture analysis;r=davehunt
We seperate logic in get-metric-for-build and update-dashboard for setting
different thresholds for startup, input, and animation analysis. This caused
things to sometimes break when an underlying component changed and one program
was modified but not the other. Fix this by consolidating all the image analysis
code inside the "eideticker" module (we'd already started to do this).

Also, we used the same language for the pixel threshold when creating the raw
framediffs as we did for getting statistics from the framediffs, which
was pretty confusing. Fix this by instead setting an option to this function
to filter out small changes (which is all ever did anyway).